siguiente commit

+++ src/main/kotlin/com/zonav/Proyectopiensa/Model/Red_point.kt
@@ -9,8 +9,8 @@ class Red_point{
     @Id
     @Column(updatable = false)
     var id: Long? = null
-    var latitude: Long? = null
-    var longitude: Long?= null
+    var latitude: Double? = null
+    var longitude: Double?= null
     var information: String?=null
     @Column(name="device_id")
     var deviceId: String?=null
