+++ src/main/kotlin/com/zonav/Proyectopiensa/Model/Yellow_point.kt
@@ -9,7 +9,7 @@ class Yellow_point{
     @Id
     @Column(updatable = false)
     var id: Long? = null
-    var latitude: Long? = null
-    var longitude: Long?= null
+    var latitude: Double? = null
+    var longitude: Double?= null
     var information: String?=null
 }
